requirements.txt
streamlit
google-api-core==1.22.1 
google-auth==1.20.1
firebase_admin
google-cloud-firestore
numpy
joblib==1.3.2
scikit-learn==1.2.2
